% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bootstrap_diffanal}
\alias{bootstrap_diffanal}
\title{Bootstrap for differential connectivity analysis (pooled bootstrap)}
\usage{
bootstrap_diffanal(df, group_col = "phenotype", max_tries = 100, seed = NULL)
}
\arguments{
\item{df}{An input n x (p+1) data frame.}

\item{group_col}{Column that represents the group/phenotype.}

\item{max_tries}{Max retries to ensure both classes appear in the bootstrap (default 100).}

\item{seed}{Optional random seed.}
}
\value{
A named list of two bootstrapped data frames (one per group),
        with \code{group_col} and the temporary \code{group_boot} removed.
}
\description{
Bootstrap for differential connectivity analysis (pooled bootstrap)
}
