% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_all}
\alias{capture_all}
\title{Silently evaluate an expression while returning its value}
\usage{
capture_all(exprs, file = NULL)
}
\arguments{
\item{exprs}{Expression to be evaluated.}

\item{file}{Optional path to a file where captured output/messages should
be written. If \code{NULL}, they are discarded to the OS null device.}
}
\value{
The normal return value of \code{exprs}.
}
\description{
Evaluates an expression, suppressing all console output
(print, cat, implicit printing) and all messages, while still
returning the normal value of the expression.
}
