% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{Compute essential centrality of a network}
\usage{
centrality(ig, weights = NULL)
}
\arguments{
\item{ig}{an igraph object}

\item{weights}{either a numeric edge-weight vector (length = |E|) or
the name of an edge attribute to use as weights (character scalar). Use NULL for unweighted.}
}
\value{
a data frame with rows = nodes and columns = centrality scores
}
\description{
Compute essential centrality of a network
}
