% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorize}
\alias{colorize}
\title{Colorize numerical values}
\usage{
colorize(values, resolution = 4)
}
\arguments{
\item{values}{values to be colorized}

\item{resolution}{Limit resolution for small values}
}
\value{
Colorized values
}
\description{
Colorize numerical values
}
