% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorize_community}
\alias{colorize_community}
\title{Assign unique color to each community}
\usage{
colorize_community(ig, community_attr_name = "community")
}
\arguments{
\item{ig}{an igraph object}

\item{community_attr_name}{the node attribute that represents the community partition}
}
\value{
an igraph object with "color" node attribute
}
\description{
Assign unique color to each community
}
