% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{community_detection}
\alias{community_detection}
\title{Perform community detection on a network}
\usage{
community_detection(
  ig,
  mode = c("total", "in", "out"),
  vertex_symbol = "name",
  method = c("walk_trap", "louvain", "leiden"),
  small_community = NULL,
  steps = 10,
  resolution = 1,
  n_iterations = 5
)
}
\arguments{
\item{ig}{an igraph object}

\item{mode}{the connection mode used to defined "isolated" nodes}

\item{vertex_symbol}{the node attribute that represents the symbol of the node}

\item{method}{the community detection method}

\item{small_community}{the threshold of small communities to be removed, is NULL, small communities will not be removed}

\item{steps}{the number of steps in walk-trap algorithm}

\item{resolution}{the resolution in Louvain and/or Leiden algorithm}

\item{n_iterations}{the number of iteration in Leiden algorithm}
}
\value{
a list object
}
\description{
Perform community detection on a network
}
