% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_net_cgbn.R
\name{consensus_net_cgbn}
\alias{consensus_net_cgbn}
\title{Create consensus network of resampling-based conditional Gaussian Bayesian network}
\usage{
consensus_net_cgbn(
  networks,
  reference_network = NULL,
  method = c("all", "average"),
  cut = 0.5
)
}
\arguments{
\item{networks}{a list of igraph directed graphs (same vertex set by name)}

\item{reference_network}{an igraph directed graph (optional, used when method = "all")}

\item{method}{the method to construct the consensus network: "all" or "average"}

\item{cut}{minimum frequency (in [0,1]) to retain an edge in the consensus}
}
\value{
a list with consensus_network (igraph) and method
}
\description{
Create consensus network of resampling-based conditional Gaussian Bayesian network
}
