% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_net_ggm.R
\name{consensus_net_ggm}
\alias{consensus_net_ggm}
\title{Create consensus network of resampling-based Markov network}
\usage{
consensus_net_ggm(
  ggm_networks,
  CI = 0.95,
  pos_cut = 0,
  neg_cut = 0,
  node_annot = NULL,
  filter = c("pval", "fdr", "none"),
  threshold = 0.05
)
}
\arguments{
\item{ggm_networks}{the output from "ensemble_sggm.R"}

\item{CI}{confidence interval of interests}

\item{pos_cut}{a threshold for positive partial correlation}

\item{neg_cut}{a threshold for negative partial correlation}

\item{node_annot}{a data frame have at least two columns: 'id' and 'symbol', and 'id' must match the colnames in "mat"}

\item{filter}{filter method}

\item{threshold}{threshold of the selected filter}
}
\value{
a list object including an igraph object as the consensus network
}
\description{
Create consensus network of resampling-based Markov network
}
