% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_net_ggm.R
\name{.abs_pcor_filter}
\alias{.abs_pcor_filter}
\title{Filter edges whose effect size is small}
\usage{
.abs_pcor_filter(x, pos_cut, neg_cut)
}
\arguments{
\item{x}{the average partial correlation}

\item{pos_cut}{a threshold for positive partial correlation}

\item{neg_cut}{a threshold for negative partial correlation}
}
\value{
the average value or 0
}
\description{
Filter edges whose effect size is small
}
\keyword{internal}
