% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_net_ggm.R
\name{.CI_filter}
\alias{.CI_filter}
\title{Filter edges whose confidence interval contains 0}
\usage{
.CI_filter(x)
}
\arguments{
\item{x}{a string in the format of "low,avg,high"}
}
\value{
the average value or 0
}
\description{
Filter edges whose confidence interval contains 0
}
\keyword{internal}
