% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_gdvm_gcm.R
\name{.count_2star_orbits}
\alias{.count_2star_orbits}
\title{Function to count 2-star orbits}
\usage{
.count_2star_orbits(edge_list, GOM)
}
\arguments{
\item{edge_list}{a matrix with 3 columns, the third column = "sign"}

\item{GOM}{the graphlet orbit matrix}
}
\value{
an updated GOM
}
\description{
Function to count 2-star orbits
}
\keyword{internal}
