% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_gdvm_gcm.R
\name{.count_tri_orbits}
\alias{.count_tri_orbits}
\title{Function to count triangle orbits}
\usage{
.count_tri_orbits(v, u, w, edge_list, GOM)
}
\arguments{
\item{v}{a vertex id}

\item{u}{a vertex id}

\item{w}{a vertex id}

\item{edge_list}{a matrix with 3 columns, the third column = "sign"}

\item{GOM}{the graphlet orbit matrix}
}
\value{
the updated graphlet orbit matrix
}
\description{
Function to count triangle orbits
}
\keyword{internal}
