% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_ggm.R
\name{.diffanal_wrapper}
\alias{.diffanal_wrapper}
\title{Run SILGGM and build an igraph with edge attributes}
\usage{
.diffanal_wrapper(
  dat,
  method,
  alpha = 0.05,
  global = TRUE,
  filter = c("pval", "fdr", "none"),
  threshold = 0.05
)
}
\arguments{
\item{dat}{n x p numeric matrix/data.frame (rows = samples, cols = features)}

\item{method}{character; inference method}

\item{alpha}{numeric;}

\item{global}{logical;}

\item{filter}{one of \code{c("pval","fdr","none")};}

\item{threshold}{numeric; threshold for edge inclusion.}
}
\value{
igraph object with edge attributes: \code{qval}, \code{pval}, \code{pcor}, \code{sign}
}
\description{
Run SILGGM and build an igraph with edge attributes
}
\keyword{internal}
