% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cgbn.R
\name{.fit_cgbn}
\alias{.fit_cgbn}
\title{Learn a conditional Gaussian Bayesian network using PC-algorithm}
\usage{
.fit_cgbn(
  dat,
  discrete_variable = NULL,
  constraints = NULL,
  alpha = 0.01,
  tol = 1e-04,
  maxit = 0
)
}
\description{
Learn a conditional Gaussian Bayesian network using PC-algorithm
}
\keyword{internal}
