% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cgbn.R
\name{.get_continuous_marginal}
\alias{.get_continuous_marginal}
\title{Obtain continuous marginals}
\usage{
.get_continuous_marginal(network, node_df)
}
\description{
Obtain continuous marginals
}
\keyword{internal}
