% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cgbn.R
\name{.get_discrete_marginal}
\alias{.get_discrete_marginal}
\title{Obtain discrete marginals}
\usage{
.get_discrete_marginal(network, node_df)
}
\description{
Obtain discrete marginals
}
\keyword{internal}
