% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_net_ggm.R
\name{.get_info}
\alias{.get_info}
\title{Get the lower or upper quantile of an edge}
\usage{
.get_info(edge, info)
}
\arguments{
\item{edge}{an edge of format (source, target)}

\item{info}{the information of interests}
}
\value{
the quantile information of the edge
}
\description{
Get the lower or upper quantile of an edge
}
\keyword{internal}
