% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cgbn.R
\name{.hugin_to_adjmat}
\alias{.hugin_to_adjmat}
\title{Coerce an RHugin graph to an adjacency matrix}
\usage{
.hugin_to_adjmat(network)
}
\description{
Coerce an RHugin graph to an adjacency matrix
}
\keyword{internal}
