% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{.leiden}
\alias{.leiden}
\title{Perform Leiden community detection algorithm on a network}
\usage{
.leiden(
  igs,
  vertex_symbol = "name",
  resolution = 1,
  objective_function = "modularity",
  n_iterations = 5
)
}
\arguments{
\item{igs}{a list containing an igraph object named "ig"}

\item{vertex_symbol}{the node attribute that represents the symbol of the node}

\item{resolution}{resolution in Leiden algorithm}

\item{objective_function}{the objective function in Leiden algorithm}

\item{n_iterations}{number of iteration in Leiden algorithm}
}
\value{
a list object
}
\description{
Perform Leiden community detection algorithm on a network
}
\keyword{internal}
