% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{.louvain}
\alias{.louvain}
\title{Perform Louvain community detection algorithm on a network}
\usage{
.louvain(igs, vertex_symbol = "name", resolution = 1)
}
\arguments{
\item{igs}{a list containing an igraph object named "ig"}

\item{vertex_symbol}{the node attribute that represents the symbol of the node}

\item{resolution}{resolution in Louvain algorithm}
}
\value{
a list object
}
\description{
Perform Louvain community detection algorithm on a network
}
\keyword{internal}
