% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_ggm.R
\name{.pcor_zscore}
\alias{.pcor_zscore}
\title{Compute the z-score of a partial correlation, see https://github.com/cran/SILGGM/blob/master/src/SILGGMCpp.cpp}
\usage{
.pcor_zscore(pcor, n)
}
\arguments{
\item{pcor}{the partial correlation}

\item{n}{the number of samples}
}
\value{
the z-score of the corresponding partial correlation
}
\description{
Compute the z-score of a partial correlation, see https://github.com/cran/SILGGM/blob/master/src/SILGGMCpp.cpp
}
\keyword{internal}
