% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_ggm.R
\name{.pow}
\alias{.pow}
\title{Raise of number to its n-th power}
\usage{
.pow(x, n)
}
\arguments{
\item{x}{the input number}

\item{n}{the order of power}
}
\value{
a number
}
\description{
Raise of number to its n-th power
}
\keyword{internal}
