% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_ggm.R
\name{.sggm_wrapper}
\alias{.sggm_wrapper}
\title{A Wrapper function of SILGGM}
\usage{
.sggm_wrapper(mat, method, alpha = 0.05, global = TRUE)
}
\arguments{
\item{mat}{a n x p matrix, n = number of samples, and p = number of features}

\item{method}{the inference method}

\item{alpha}{A user-supplied sequence of pre-sepecified alpha levels for FDR control.}

\item{global}{glob == FALSE -> only positive correlation; glob == TRUE -> ± partial correlations}
}
\value{
a list object containing two matrics
}
\description{
A Wrapper function of SILGGM
}
\keyword{internal}
