% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_net_ggm.R
\name{.to_numeric}
\alias{.to_numeric}
\title{Convert the string of "low,avg,high" to numeric value}
\usage{
.to_numeric(x)
}
\arguments{
\item{x}{a string in the format of "low,avg,high"}
}
\value{
a numeric vector
}
\description{
Convert the string of "low,avg,high" to numeric value
}
\keyword{internal}
