% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_detection.R
\name{.walktrap}
\alias{.walktrap}
\title{Perform walk-trap community detection algorithm on a network}
\usage{
.walktrap(igs, vertex_symbol = "name", steps = 10)
}
\arguments{
\item{igs}{a list containing an igraph object named "ig"}

\item{vertex_symbol}{the node attribute that represents the symbol of the node}

\item{steps}{the number of steps in walk-trap algorithm}
}
\value{
a list object
}
\description{
Perform walk-trap community detection algorithm on a network
}
\keyword{internal}
