% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_cgbn.R
\name{ensemble_cgbn}
\alias{ensemble_cgbn}
\title{Resampling-based conditional Gaussian Bayesian Network}
\usage{
ensemble_cgbn(
  dat,
  discrete_variable = NULL,
  num_iteration = 1,
  boot = FALSE,
  sub_ratio = 0.9,
  sample_class = NULL,
  hugin = FALSE,
  constraints = NULL,
  alpha = 0.01,
  tol = 1e-04,
  maxit = 0,
  n_cores = NULL
)
}
\arguments{
\item{dat}{a n x p data frame/matrix with row and column names, n = number of samples, and p = number of features}

\item{discrete_variable}{the columns that represents the discrete_variable, must be a character vector}

\item{num_iteration}{an integer, indicating the number of iteration of resampling}

\item{boot}{a logical variable, if TRUE, then perform bootstrap resampling, else, perform subsampling}

\item{sub_ratio}{a numerical value between 0 and 1, indicating the subsampling ratio}

\item{sample_class}{a atomic vector, indicating the class of each sample, if != NULL, then stratified sampling is performed}

\item{hugin}{a logical variable, if TRUE, the Hugin object will be included in the output}

\item{constraints}{prior structure knowledge, default = NULL, details -> RHugin::learn.structure}

\item{alpha}{parameter of PC algorithm for structure learning}

\item{tol}{parameter of EM algorithm for CPT learning, details -> RHugin::learn.cpt}

\item{maxit}{parameter of EM algorithm for CPT learning, details -> RHugin::learn.cpt}

\item{n_cores}{number of cores for parallel computing}
}
\value{
a list object
}
\description{
Resampling-based conditional Gaussian Bayesian Network
}
