% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean_pairwise_complete_obs}
\alias{euclidean_pairwise_complete_obs}
\title{Function to compute Euclidean distance with "pairwise.complete.obs" method}
\usage{
euclidean_pairwise_complete_obs(v1, v2)
}
\arguments{
\item{v1}{the first vector}

\item{v2}{the second vector}
}
\value{
the Euclidean distance
}
\description{
Function to compute Euclidean distance with "pairwise.complete.obs" method
}
