% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_distance.R
\name{gcm_distance}
\alias{gcm_distance}
\title{Distance between two Graphlet Correlation Matrices}
\usage{
gcm_distance(gcm1, gcm2)
}
\arguments{
\item{gcm1}{the GCM of a network}

\item{gcm2}{the GCM of another network}
}
\value{
numeric; the Euclidean distance between the upper triangular of gcm1 and gcm2
}
\description{
Distance between two Graphlet Correlation Matrices
}
