% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdv_distance.R
\name{gdv_distance}
\alias{gdv_distance}
\title{Find the graphlet degree distance between two vertices}
\usage{
gdv_distance(v_gdv, u_gdv, w = 1, similarity = FALSE)
}
\arguments{
\item{v_gdv}{the graphlet degree vector of the first node, denoted as v}

\item{u_gdv}{the graphlet degree vector of the second node, denoted as u}

\item{w}{the weight vector assigned to each orbit, w = 1 for all orbits by default}

\item{similarity}{if True, return the similarity score, and distance otherwise}
}
\value{
the distance or similarity score
}
\description{
Find the graphlet degree distance between two vertices
}
\references{
Milenković, Tijana, and Nataša Pržulj. "Uncovering biological network function via graphlet degree signatures."
}
