% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdvm_gcm.R
\name{gdvm_gcm}
\alias{gdvm_gcm}
\title{Find graphlet degree vector (GDV) of each vertex and graphlet correlation matrix of the given network}
\usage{
gdvm_gcm(
  ig,
  level = c("4", "5"),
  redundant_orbit = FALSE,
  include_gcm = FALSE,
  correlation = "spearman"
)
}
\arguments{
\item{ig}{an igraph object}

\item{level}{node-level of graphlet}

\item{redundant_orbit}{if TRUE, keep the redundant orbits, and remove otherwise}

\item{include_gcm}{if TRUE, compute the GCM}

\item{correlation}{the correlation measure between each orbit, "spearman" correlation is applied by default}
}
\value{
a list of two matrics, gdvm = graphlet degree vector matrix, gcm = graphlet correlation matrix
}
\description{
Find graphlet degree vector (GDV) of each vertex and graphlet correlation matrix of the given network
}
\references{
Yaveroğlu, Ömer Nebil, et al. "Revealing the hidden language of complex networks."
}
