% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Calculate the Jaccard similarity between vectors or matrices}
\usage{
jaccard_similarity(x, y)
}
\arguments{
\item{x}{a vector or a adjacency matrix}

\item{y}{a vector or a adjacency matrix}
}
\value{
the Jaccard similarity between x and y
}
\description{
Calculate the Jaccard similarity between vectors or matrices
}
