% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_er}
\alias{model_er}
\title{Simulate network via erdos-renyi model}
\usage{
model_er(p = 300, e_prob = 0.01, DAG = FALSE)
}
\arguments{
\item{p}{The number of vertices in the graph}

\item{e_prob}{A probability for drawing an edge between two arbitrary vertices}

\item{DAG}{If TRUE, simulate an directed acyclic graph, or DAG}
}
\value{
An igraph object
}
\description{
Simulate network via erdos-renyi model
}
