% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_mpa}
\alias{model_mpa}
\title{Simulate network via modular preferential attachment model}
\usage{
model_mpa(p = 300, m = 6, q_hub = 0.95, m_links = 6, power = 1.7, z_appeal = 1)
}
\arguments{
\item{p}{The number of vertices in the graph}

\item{m}{The number of modules}

\item{q_hub}{The quantile for degree cutoff for defining modular hubs}

\item{m_links}{The number of links between modules}

\item{power}{The power of the preferential attachment}

\item{z_appeal}{The attractiveness of the vertices with no adjacent edges}
}
\value{
An igraph object
}
\description{
Simulate network via modular preferential attachment model
}
