% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_pa}
\alias{model_pa}
\title{Simulate network via preferential attachment model}
\usage{
model_pa(p = 300, power = 1, z_appeal = 1, DAG = FALSE)
}
\arguments{
\item{p}{The number of vertices in the graph}

\item{power}{The power of the preferential attachment}

\item{z_appeal}{The attractiveness of the vertices with no adjacent edges}

\item{DAG}{If TRUE, simulate an directed acyclic graph, or DAG}
}
\value{
An igraph object
}
\description{
Simulate network via preferential attachment model
}
