% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_sbm}
\alias{model_sbm}
\title{Simulate network via stochastic block model}
\usage{
model_sbm(p = 300, m = 6, sizes = NULL, p_in = 0.8, p_out = 0.05)
}
\arguments{
\item{p}{The number of vertices in the graph}

\item{m}{The number of modules}

\item{sizes}{Size of each module, default = equal-sized modules}

\item{p_in}{within-module edge density}

\item{p_out}{between-module edges}
}
\value{
An igraph object
}
\description{
Simulate network via stochastic block model
}
