% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_sim}
\alias{model_sim}
\title{Simulate multivariate Gaussian data for an undirected graph model}
\usage{
model_sim(n, ig, seed = 1)
}
\arguments{
\item{n}{The number of samples to generate data for}

\item{ig}{An igraph object, must be an undirected graph}

\item{seed}{A number to seed the random data}
}
\value{
A data frame of multivariate Gaussian data
}
\description{
Simulate multivariate Gaussian data for an undirected graph model
}
