% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_sw}
\alias{model_sw}
\title{Simulate network via small world model}
\usage{
model_sw(p = 300, dim = 1, nei = 3, rw_prob = 0.05)
}
\arguments{
\item{p}{The number of vertices in the graph}

\item{dim}{The dimension of the starting lattice}

\item{nei}{The neighborhood within which the vertices of the lattice will be connected}

\item{rw_prob}{The probability for rewiring}
}
\value{
An igraph object
}
\description{
Simulate network via small world model
}
