% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permutation_diffanal}
\alias{permutation_diffanal}
\title{Permutation for differential connectivity analysis}
\usage{
permutation_diffanal(
  df,
  group_col = "phenotype",
  balanced = FALSE,
  seed = NULL
)
}
\arguments{
\item{df}{an input n by (p+1) data frame}

\item{group_col}{the column that represents the "group/phenotype" to be shuffled}

\item{balanced}{a logical parameter for regular or balanced permutation}

\item{seed}{random seed}
}
\value{
an list object containing two shuffled data frame for each group
}
\description{
Permutation for differential connectivity analysis
}
