% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pgdv_distance}
\alias{pgdv_distance}
\title{Pair-wise GDV distance}
\usage{
pgdv_distance(gdvm, w = 1, similarity = FALSE)
}
\arguments{
\item{gdvm}{the graphlet degree vector matrix}

\item{w}{the weight vector}

\item{similarity}{if True, return the similarity score, and distance otherwise}
}
\value{
a n x n distance/similarity matrix, n = # of node
}
\description{
Pair-wise GDV distance
}
