% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_isolated}
\alias{remove_isolated}
\title{Remove isolated nodes from a network}
\usage{
remove_isolated(ig, mode = c("total", "in", "out"), vertex_symbol = NULL)
}
\arguments{
\item{ig}{a igraph object}

\item{mode}{the connection mode}

\item{vertex_symbol}{the node attribute that represents the symbol of the node}
}
\value{
a list object
}
\description{
Remove isolated nodes from a network
}
