% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_small_community}
\alias{remove_small_community}
\title{Remove small communities in a network}
\usage{
remove_small_community(igs, threshold = 5, community = "community")
}
\arguments{
\item{igs}{a list containing at least: an igraph object with name "ig", and "ig" must have vertex attribute of community partition}

\item{threshold}{the minimum community size}

\item{community}{the name of vertex attribute "community"}
}
\value{
a list object
}
\description{
Remove small communities in a network
}
