% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample}
\alias{resample}
\title{Perform data resampling}
\usage{
resample(sample_data, sample_class, boot = TRUE, sub_ratio = NULL)
}
\arguments{
\item{sample_data}{a n x 2 data frame is "sample_class" is provided, else, n x 1 data frame of sample IDs}

\item{sample_class}{a vector indicates the class of each sample, if != NULL, then stratified sampling is performed}

\item{boot}{a logical variable, if TRUE, then perform bootstrap resampling}

\item{sub_ratio}{a numerical value between 0 and 1, indicating the subsampling ratio, or NULL}
}
\value{
a vector containing the resampled row.names(sample_data)
}
\description{
Perform data resampling
}
