% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample_cluster}
\alias{resample_cluster}
\title{Perform data resampling from correlated data}
\usage{
resample_cluster(
  sample_data,
  id_col = "ID",
  cluster_col = "cluster",
  boot = TRUE,
  ratio = NULL
)
}
\arguments{
\item{sample_data}{a data frame of sample meta-data}

\item{id_col}{the column name specified the unique sample identifiers}

\item{cluster_col}{the column name represents the "cluster" of each sample}

\item{boot}{a logical variable, if TRUE, then perform bootstrap resampling}

\item{ratio}{the ratio of clusters subject to bootstrapping}
}
\value{
a list object containing bootstrapping cluster and associated samples
}
\description{
Perform data resampling from correlated data
}
