% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed_gdvm_gcm.R
\name{signed_gdvm_gcm}
\alias{signed_gdvm_gcm}
\title{Function to count signed graphlets}
\usage{
signed_gdvm_gcm(ig, n_cores = NULL, redundant = FALSE, include_gcm = FALSE)
}
\arguments{
\item{ig}{an igraph object}

\item{n_cores}{number of cores for parrallel computing}

\item{redundant}{if TRUE, include the redundant orbits}

\item{include_gcm}{if TRUE, compute GCM}
}
\value{
a list of two matrics, gdvm = graphlet degree vector matrix, gcm = graphlet correlation matrix
}
\description{
Function to count signed graphlets
}
\references{
Das, Apratim. Efficient enumeration of small graphlets and orbits.
}
