% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{sim_fam_data}
\alias{sim_fam_data}
\title{Simulate family-based correlated data}
\usage{
sim_fam_data(num_families = 100, members_per_family = 5, h_sq = 0.5, Sigma)
}
\arguments{
\item{num_families}{number of family}

\item{members_per_family}{number of individuals in each family}

\item{h_sq}{inheritability}

\item{Sigma}{the covariance matrix of variables}
}
\value{
an ExpressionSet object
}
\description{
Simulate family-based correlated data
}
