% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_cgbn}
\alias{toy_cgbn}
\title{Toy dataset of mixture data type}
\format{
A data frame
}
\usage{
data(toy_cgbn)
}
\value{
A data frame
}
\description{
A 50 x 25 data frame. with 20 continuous and 5 discrete variables
}
\keyword{datasets}
