% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_er}
\alias{toy_er}
\title{Toy undirected network generated by ER-model}
\format{
A list object of an igraph obejct and a data frame
}
\usage{
data(toy_er)
}
\value{
A list object
}
\description{
A list of an igraph object and a data frame
}
\keyword{datasets}
