% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_load}
\alias{toy_load}
\title{Toy dataset from LOAD study}
\format{
A data frame object
}
\usage{
data(toy_load)
}
\value{
A data frame object
}
\description{
A data frame containing LOAD and Control subjects
}
\references{
Zhang, Bin, et al.
"Integrated systems approach identifies genetic nodes and networks in late-onset Alzheimer’s disease."
Cell 153.3 (2013): 707-720.
}
\keyword{datasets}
