% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upper_tri_to_matrix}
\alias{upper_tri_to_matrix}
\title{Reconstruct the symmetric matrix from upper triangular vector}
\usage{
upper_tri_to_matrix(upper_tri_values, variable_names = NULL, diagl = 1)
}
\arguments{
\item{upper_tri_values}{a numeric vector of the upper triangle of the matrix}

\item{variable_names}{row&column names}

\item{diagl}{the diagnoal elemenet, a number of a numerci vector}
}
\value{
a symmetric matrix
}
\description{
Reconstruct the symmetric matrix from upper triangular vector
}
