% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cn.R
\name{plot_cn}
\alias{plot_cn}
\title{Plot Consensus Network}
\usage{
plot_cn(
  ig,
  query = NULL,
  order = 1,
  vertex_symbol = "symbol",
  edge_label = NULL,
  edge_width = NULL,
  CI_show = FALSE,
  main = "Title",
  query_color = "#FF9999",
  neighbor_color = "#56B4E9",
  all_color = "#56B4E9",
  hover = TRUE,
  stabilization = FALSE,
  igraph_layout = "layout_with_fr",
  layout_seed = 42
)
}
\arguments{
\item{ig}{an igraph object}

\item{query}{query nodes for subgraph}

\item{order}{the order of neighbors of the queried nodes}

\item{vertex_symbol}{the node attribute that represents the symbol}

\item{edge_label}{the edge attributes that represents the edge label}

\item{edge_width}{the edge attributes that represents the edge width}

\item{CI_show}{if TRUE, plot the confidence interval}

\item{main}{title}

\item{query_color}{the color of queried nodes}

\item{neighbor_color}{the color of the neighbors of queries nodes}

\item{all_color}{the color for all nodes in the network}

\item{hover}{a logical argument}

\item{stabilization}{a logical argument}

\item{igraph_layout}{the layout being used}

\item{layout_seed}{the random seed of the layout}
}
\value{
a list containing a visNetwork object
}
\description{
Plot Consensus Network
}
